Fix HOH coorinating metals in PLEC

+++ oddt/toolkits/ob.py
@@ -379,7 +379,7 @@ class Molecule(pybel.Molecule):
     def _dicts(self):
         max_neighbors = 6  # max of 6 neighbors should be enough
         # Atoms
-        atom_dtype = [('id', np.uint16),
+        atom_dtype = [('id', np.uint32),
                       # atom info
                       ('coords', np.float32, 3),
                       ('radius', np.float32),
